Complete refactor
- Base functionality extracted completely
- Game aspects separated into appropriate modules
- Single Responsibility intact
