+++ lib/index.js
@@ -1,3 +1,13 @@
-var Game = require("./game");
+var KeyListener = require("./keyListener");
+var Game        = require("./game");
+var Render      = require("./render");
 
-new Game();
+KeyListener.start();
+var game = new Game();
+requestAnimationFrame(gameLoop);
+
+function gameLoop(){
+  game.update();
+  Render(game.firstPlayer);
+  requestAnimationFrame(gameLoop);
+}
